Changed CSS navigation system.
Utter change on how components are shown according to the user
navigation.

The navigation system should rely on a sole index.html, in order to
prevent reload index.js (and fetch a huge data amount) on topic
changing.

In consequence, the structure which represents these topics are
sectioned by unique segments that should be displayed according
to a anchor which navigates to and hides all the others.

Highlights:
* The heading structure was rewritten into three levels in index.html,
  in order to accommodate a unique higher level heading;
* Style for pagination was created;
* Style for header was utterly changed to favor navigation items'
  responsivity;
* Removed 'no-content' block since it is no longer necessary;
* Other .html files were removed.